% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/_DataStorageClass.R
\docType{class}
\name{DatNetStorageClass}
\alias{DatNetStorageClass}
\title{R6 class for storing, managing, subsetting and manipulating the input data for networks.}
\format{An \code{\link{R6Class}} generator object}
\usage{
DatNetStorageClass
}
\description{
R6 class for storing, managing, subsetting and manipulating the input data for networks.
}
\details{
#' \itemize{
\item{\code{sW}} - Baseline summaries
\item{\code{sA}} - Exposure summaries
\item{\code{intervene1.sA}} - Intervention object 1
\item{\code{intervene2.sA}} - Intervention object 2
}
}
\section{Methods}{

\describe{
  \item{\code{new(Odata, nodes, YnodeVals, det.Y, ...)}}{...}
  \item{\code{make.sVar(type.sVar = NULL)}}{...}
}
}

\section{Active Bindings}{

\describe{
   \item{\code{save_sA_Vars}}{...}
   \item{\code{restored_sA_Vars}}{...}
}
}
\keyword{R6}
\keyword{class}

