% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Learner_syntax.R
\name{defModel}
\alias{defModel}
\alias{+.ModelStack}
\title{Interface for defining models}
\usage{
defModel(estimator, x, search_criteria, param_grid, ...)

\method{+}{ModelStack}(model1, model2)
}
\arguments{
\item{estimator}{A character string name of package and estimator (algorithm) name, separated by "__".}

\item{x}{A vector containing the subset of the names of the predictor variables to use in building this
particular learner or grid. This argument can be used to over-ride the values of \code{x} provided to \code{fit} function.
As such, the names supplied here must always be a subset of the names specified to \code{fit}.
When this argument is missing (default) the column names provided to \code{fit} are used as predictors in building this model / grid.}

\item{search_criteria}{Search criteria}

\item{param_grid}{Named list of model hyper parameters (optional).
Each named item defines either a fixed model parameter value or a vector of possible values.
In the latter case this function call would define a grid (ensemble) of models.
Each model in the ensemble corresponds by a particular fixed parameter value.}

\item{...}{Additional modeling parameters to be passed on directly to the modeling function.}

\item{model1}{An object returned by a call to \code{defModel} function.}

\item{model2}{An object returned by a call to \code{defModel} function.}
}
\description{
Specify either a single model or a grid of multple models by invoking the optional argument \code{param_grid}.
}
