% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stremr-package.R
\docType{data}
\name{dt_1t}
\alias{dt_1t}
\title{An example of a dataset in long format with one time-point and no censoring.}
\format{
A data frame with 50,000 observations and variables:
\describe{
  \item{ID}{Unique subject identifier}
  \item{L1}{Binary baseline confounder}
  \item{L2}{Binary baseline confounder}
  \item{L3}{Binary baseline confounder}
  \item{L4}{Binary baseline confounder}
  \item{L5}{Binary baseline confounder}
  \item{L6}{Binary baseline confounder}
  \item{A}{Binary exposure}
  \item{Y}{Indicator of the binary outcome at the end of the study.}
}
}
\usage{
data(dt_1t)
}
\description{
Simulated dataset containing 50,000 i.i.d. observations with single time-point exposure.
The binary exposure is \code{A} and binary outcome is \code{Y}. See /tests/
for R code that generated this data as well as R code that uses stremr to analyze this data.
}
\keyword{datasets}
