% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stremr-package.R
\docType{data}
\name{dt_2t}
\alias{dt_2t}
\title{An example of a dataset in long format with two time-points and no censoring.}
\format{
A data frame with 500,000 observations and variables:
\describe{
  \item{ID}{Unique subject identifier}
  \item{t}{Interger for current time period, range 0-1}
  \item{L1}{Binary time-varying confounder}
  \item{L2}{Binary time-varying confounder}
  \item{L3}{Binary time-varying confounder}
  \item{A}{Binary time-varying exposure}
  \item{Y}{Indicator of the binary outcome at the end of the study (t=1). Constant set to 0 at first time point (t=0)}
}
}
\usage{
data(dt_2t)
}
\description{
Simulated dataset containing 500,000 i.i.d. observations organized in long format as person-time rows.
The binary time-varying exposure is \code{A} and binary time-invariant outcome is \code{Y}. See /tests/
for R code that generated this data as well as R code that uses stremr to analyze this data.
}
\keyword{datasets}
