% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmle_SDR_iTMLE.R
\name{fit_iTMLE}
\alias{fit_iTMLE}
\title{Fit the sequential double robust (SDR) procedure, either with DR tranformation or TMLE-like targeting}
\usage{
fit_iTMLE(
  OData,
  tvals,
  Qforms,
  intervened_TRT = NULL,
  intervened_MONITOR = NULL,
  rule_name = paste0(c(intervened_TRT, intervened_MONITOR), collapse = ""),
  models = NULL,
  fit_method = stremrOptions("fit_method"),
  fold_column = stremrOptions("fold_column"),
  stratifyQ_by_rule = FALSE,
  stratify_by_last = TRUE,
  useonly_t_TRT = NULL,
  useonly_t_MONITOR = NULL,
  CVTMLE = FALSE,
  trunc_weights = 10^6,
  weights = NULL,
  parallel = FALSE,
  return_fW = FALSE,
  use_DR_transform = FALSE,
  stabilize = FALSE,
  reg_Q = NULL,
  SDR_model = NULL,
  verbose = getOption("stremr.verbose"),
  ...
)
}
\arguments{
\item{OData}{Input data object created by \code{importData} function.}

\item{tvals}{Vector of time-points in the data for which the survival function (and risk) should be estimated}

\item{Qforms}{Regression formulas, one formula per Q. Only main-terms are allowed.}

\item{intervened_TRT}{Column name in the input data with the probabilities (or indicators) of counterfactual treatment nodes being equal to 1 at each time point.
Leave the argument unspecified (\code{NULL}) when not intervening on treatment node(s).}

\item{intervened_MONITOR}{Column name in the input data with probabilities (or indicators) of counterfactual monitoring nodes being equal to 1 at each time point.
Leave the argument unspecified (\code{NULL}) when not intervening on the monitoring node(s).}

\item{rule_name}{Optional name for the treatment/monitoring regimen.}

\item{models}{Optional parameters specifying the models for fitting the iterative (sequential) G-Computation formula.
Must be an object of class \code{ModelStack} specified with \code{gridisl::defModel} function.}

\item{fit_method}{Model selection approach. Can be either \code{"none"} - no model selection or
\code{"cv"} - V fold cross-validation that selects the best model according to lowest cross-validated MSE (must specify the column name that contains the fold IDs).}

\item{fold_column}{The column name in the input data (ordered factor) that contains the fold IDs to be used as part of the validation sample.
Use the provided function \code{\link{define_CVfolds}} to
define such folds or define the folds using your own method.}

\item{stratifyQ_by_rule}{Set to \code{TRUE} for stratifying the fit of Q (the outcome model) by rule-followers only.
There are two ways to do this stratification. The first option is to use \code{stratify_by_last=TRUE}  (default),
which would fit the outcome model only among the observations that were receiving their supposed
counterfactual treatment at the current time-point (ignoring the past history of treatments leading up to time-point t).
The second option is to set \code{stratify_by_last=FALSE} in which case the outcome model will be fit only
among the observations who followed their counterfactual treatment regimen throughout the entire treatment history up to
current time-point t (rule followers). For the latter option, the observation would be considered a non-follower if
the person's treatment did not match their supposed counterfactual treatment at any time-point up to and including current
time-point t.}

\item{stratify_by_last}{Only used when \code{stratifyQ_by_rule} is \code{TRUE}.
Set to \code{TRUE} for stratification by last time-point, set to \code{FALSE} for stratification by all time-points (rule-followers).
See \code{stratifyQ_by_rule} for more details.}

\item{useonly_t_TRT}{Use for intervening only on some subset of observation and time-specific treatment nodes.
Should be a character string with a logical expression that defines the subset of intervention observations.
For example, using \code{TRT==0} will intervene only at observations with the value of \code{TRT} being equal to zero.
The expression can contain any variable name that was defined in the input dataset.
Leave as \code{NULL} when intervening on all observations/time-points.}

\item{useonly_t_MONITOR}{Same as \code{useonly_t_TRT}, but for monitoring nodes.}

\item{CVTMLE}{Set to \code{TRUE} to run the CV-TMLE algorithm instead of the usual TMLE algorithm.
Must set either \code{TMLE}=\code{TRUE} or \code{iterTMLE}=\code{TRUE} for this argument to have any effect..}

\item{trunc_weights}{Specify the numeric weight truncation value. All final weights exceeding the value in \code{trunc_weights} will be truncated.}

\item{weights}{Optional \code{data.table} with additional observation- and time-specific weights.  Must contain columns \code{ID}, \code{t} and \code{weight}.
The column named \code{weight} is merged back into the original data according to (\code{ID}, \code{t}). Not implemented yet.}

\item{parallel}{Set to \code{TRUE} to run the sequential G-COMP or TMLE in parallel (uses \code{foreach} with \code{dopar} and
requires a previously defined parallel back-end cluster)}

\item{return_fW}{When \code{TRUE}, will return the object fit for the last Q regression as part of the output table.
Can be used for obtaining subject-specific predictions of the counterfactual functional E(Y_{d}|W_i).}

\item{use_DR_transform}{Apply DR transform estimator instead of the iTMLE.}

\item{stabilize}{Only applies when \code{use_DR_transform=TRUE}. Set this argument to \code{TRUE} to stabilize the weights by the
empirical conditional probability of having followed the rule at time-point \code{t}, given the subject has followed the rule all
the way up to time-point \code{t}.}

\item{reg_Q}{(ADVANCED USE ONLY) Directly specify the Q regressions, separately for each time-point.}

\item{SDR_model}{The xgboost parameter settings for iTMLE non-parametric regression targeting.
If missing/NULL the default parameter settings will be used.}

\item{verbose}{Set to \code{TRUE} to print auxiliary messages during model fitting.}

\item{...}{When \code{models} arguments is NOT specified, these additional arguments will be passed on directly to all \code{GridSL}
modeling functions that are called from this routine,
e.g., \code{family = "binomial"} can be used to specify the model family.
Note that all such arguments must be named.}
}
\value{
An output list containing the \code{data.table} with survival estimates over time saved as \code{"estimates"}.
}
\description{
Interventions on up to 3 nodes are allowed: \code{CENS}, \code{TRT} and \code{MONITOR}.
Adjustment will be based on the inverse of the propensity score fits for the observed likelihood (g0.C, g0.A, g0.N),
multiplied by the indicator of not being censored and the probability of each intervention in \code{intervened_TRT} and \code{intervened_MONITOR}.
Requires column name(s) that specify the counterfactual node values or the counterfactual probabilities of each node being 1 (for stochastic interventions).
}
