% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmle.R
\name{fitIterTMLE}
\alias{fitIterTMLE}
\title{Iterative TMLE wrapper for \code{fitSeqGcomp}}
\usage{
fitIterTMLE(...)
}
\arguments{
\item{...}{Arguments that will be passed down to the underlying function \code{fitSeqGcomp}}
}
\value{
\code{data.table} with survival by time for sequential GCOMP and iterative TMLE
}
\description{
Calls \code{fitSeqGcomp} with argument \code{iterTMLE = TRUE}.
}
\seealso{
\code{\link{fitSeqGcomp}}
}

