% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_estimation.R
\name{fitPropensity}
\alias{fitPropensity}
\title{Define and fit propensity score models.}
\usage{
fitPropensity(OData, gform_CENS, gform_TRT, gform_MONITOR,
  stratify_CENS = NULL, stratify_TRT = NULL, stratify_MONITOR = NULL,
  params_CENS = list(), params_TRT = list(), params_MONITOR = list(),
  verbose = getOption("stremr.verbose"))
}
\arguments{
\item{OData}{Input data object created by \code{importData} function.}

\item{gform_CENS}{...}

\item{gform_TRT}{...}

\item{gform_MONITOR}{...}

\item{stratify_CENS}{...}

\item{stratify_TRT}{...}

\item{stratify_MONITOR}{...}

\item{params_CENS}{...}

\item{params_TRT}{...}

\item{params_MONITOR}{...}

\item{verbose}{Set to \code{TRUE} to print messages on status and information to the console. Turn this on by default using \code{options(stremr.verbose=TRUE)}.}
}
\value{
...
}
\description{
Defines and fits regression models for the propensity scores for censoring, treatment and monitoring.
}

