% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmle.R
\name{fitSeqGcomp}
\alias{fitSeqGcomp}
\title{Fit sequential GCOMP and TMLE for survival}
\usage{
fitSeqGcomp(OData, t_periods, Qforms, Qstratify = NULL,
  intervened_TRT = NULL, intervened_MONITOR = NULL, useonly_t_TRT = NULL,
  useonly_t_MONITOR = NULL, rule_name = paste0(c(intervened_TRT,
  intervened_MONITOR), collapse = ""), stratifyQ_by_rule = FALSE,
  TMLE = FALSE, iterTMLE = FALSE, IPWeights = NULL, stabilize = FALSE,
  trunc_weights = 10^6, params_Q = list(), weights = NULL,
  max_iter = 50, tol.eps = 0.001, parallel = FALSE,
  verbose = getOption("stremr.verbose"))
}
\arguments{
\item{OData}{Input data object created by \code{importData} function.}

\item{t_periods}{Specify the vector of time-points for which the survival function (and risk) should be estimated}

\item{Qforms}{Regression formulas, one formula per Q. Only main-terms are allowed.}

\item{Qstratify}{Placeholder for future user-defined model stratification for all Qs (CURRENTLY NOT FUNCTIONAL, WILL RESULT IN ERROR)}

\item{intervened_TRT}{Column name in the input data with the probabilities (or indicators) of counterfactual treatment nodes being equal to 1 at each time point.
Leave the argument unspecified (\code{NULL}) when not intervening on treatment node(s).}

\item{intervened_MONITOR}{Column name in the input data with probabilities (or indicators) of counterfactual monitoring nodes being equal to 1 at each time point.
Leave the argument unspecified (\code{NULL}) when not intervening on the monitoring node(s).}

\item{useonly_t_TRT}{Use for intervening only on some subset of observation and time-specific treatment nodes.
Should be a character string with a logical expression that defines the subset of intervention observations.
For example, using \code{TRT==0} will intervene only at observations with the value of \code{TRT} being equal to zero.
The expression can contain any variable name that was defined in the input dataset.
Leave as \code{NULL} when intervening on all observations/time-points.}

\item{useonly_t_MONITOR}{Same as \code{useonly_t_TRT}, but for monitoring nodes.}

\item{rule_name}{Optional name for the treatment/monitoring regimen.}

\item{stratifyQ_by_rule}{Set to \code{TRUE} for stratification, fits the outcome model (Q-learning) among rule-followers only.
Setting to \code{FALSE} will fit the outcome model (Q-learning) across all observations (pooled regression).}

\item{TMLE}{Set to \code{TRUE} to run the usual longitudinal TMLE algorithm (with a separate TMLE update of Q for every sequential regression).}

\item{iterTMLE}{Set to \code{TRUE} to run the iterative univariate TMLE instead of the usual longitudinal TMLE.
When set to \code{TRUE} this will also provide the standard sequential Gcomp as party of the output.
Must set \code{TMLE}=\code{FALSE} when setting this to \code{TRUE}.}

\item{IPWeights}{(Optional) result of calling function \code{getIPWeights} for running TMLE (evaluated automatically when missing)}

\item{stabilize}{Set to \code{TRUE} to use stabilized weights for the TMLE}

\item{trunc_weights}{Specify the numeric weight truncation value. All final weights exceeding the value in \code{trunc_weights} will be truncated.}

\item{params_Q}{Optional parameters to be passed to the specific fitting algorithm for Q-learning}

\item{weights}{Optional \code{data.table} with additional observation-time-specific weights.  Must contain columns \code{ID}, \code{t} and \code{weight}.
The column named \code{weight} is merged back into the original data according to (\code{ID}, \code{t}).}

\item{max_iter}{Maximum number of iterations for iterative TMLE algorithm}

\item{tol.eps}{Numeric error tolerance for the iterative TMLE update.
The iterative TMLE algorithm will stop when the absolute value of the TMLE intercept update is below \code{tol.eps}}

\item{parallel}{Set to \code{TRUE} to run the sequential Gcomp or TMLE in parallel (uses \code{foreach} with \code{dopar} and requires a previously defined parallel back-end cluster)}

\item{verbose}{...}
}
\value{
...
}
\description{
Interventions on up to 3 nodes are allowed: \code{CENS}, \code{TRT} and \code{MONITOR}.
TMLE adjustment will be based on the inverse of the propensity score fits for the observed likelihood (g0.C, g0.A, g0.N),
multiplied by the indicator of not being censored and the probability of each intervention in \code{intervened_TRT} and \code{intervened_MONITOR}.
Requires column name(s) that specify the counterfactual node values or the counterfactual probabilities of each node being 1 (for stochastic interventions).
}
\seealso{
\code{\link{stremr-package}} for the general overview of the package,
}

