% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmle.R
\name{fitTMLE}
\alias{fitTMLE}
\title{TMLE wrapper for \code{fitSeqGcomp}}
\usage{
fitTMLE(...)
}
\value{
...
}
\description{
Calls \code{fitSeqGcomp} with argument \code{TMLE = TRUE}.
}
\seealso{
\code{\link{fitSeqGcomp}}
}

