% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_routines.R
\name{follow.rule.d.DT}
\alias{follow.rule.d.DT}
\title{Define the indicators of following/not-following specific dynamic treatment rules indexed by theta.}
\usage{
follow.rule.d.DT(data, theta, ID, t, I, CENS, TRT, MONITOR, rule.names = NULL,
  return.allcolumns = FALSE)
}
\arguments{
\item{data}{Input data.frame or data.table in long format, see below for the description of the assumed format.}

\item{theta}{The vector of continuous cutoff values that index each dynamic treatment rule}

\item{ID}{The name of the unique subject identifier}

\item{t}{The name of the variable indicating time-period}

\item{I}{Continuous biomarker variable used for determining the treatmet decision rule}

\item{CENS}{Binary indicator of being censored at t;}

\item{TRT}{Binary indicator of the treatment (exposure) at t;}

\item{MONITOR}{The indicator of having a visit and having measured or observed biomarker I(t+1) (the biomarker value at THE NEXT TIME CYCLE).
In other words the value of MONITOR(t-1) (at t-1) being 1 indicates that I(t) at time point t was observed/measured.
The very first value of I(t) (at the first time-cycle) is ALWAYS ASSUMED observed/measured.}

\item{rule.names}{Vector of column names for indicators of following/not following each rule (must be the same dimension as theta).
When not supplied the following convention is adopted for naming these columns: paste0("d",theta).}

\item{return.allcolumns}{Set to \code{TRUE} to return the original data columns along with new columns that define each rule
(can be useful when employing piping/sequencing operators such as %>%).}
}
\value{
A data.table with a separate column for each value in \code{theta}. Each column consists of indicators of following/not-following
 each rule indexed by a value form \code{theta}. In addition, the returned data.table contains \code{ID} and \code{t} columns for easy merging
 with the original data.
}
\description{
Define the indicators of following/not-following specific dynamic treatment rules indexed by theta.
}
\section{Details}{


* This function takes an input data.frame or data.table data
  and produces an output data.table with indicators/probabilities of following not following a specific treatment rule.
  The resulting data.table is used internally by the stremr() function to determine which observation is following each specific rule at each time point t.

* Evaluates which observations were following the dynamic-decision treatment rule defined by the measured biomarker I
  and pre-defined cutoffs of the input vector theta.

* Produces a separate rule indicator column for each value in the input vector theta based on the following dynamic rule at t:
\itemize{
\item (1) Follow rule at t if uncensored (C(t)=0) and remaining on treatment (A(t-1)=A(t)=1)
\item (2) Follow rule at t if uncensored (C(t)=0), haven't changed treatment and wasn't monitored (MONITOR(t-1)=0)
\item (3) Follow rule at t if uncensored (C(t)=0), was monitored (MONITOR(t-1)=0) and either of the two:
    (A) (I(t) >= d.theta) and switched to treatment at t; or (B) (I(t) < d.theta) and haven't changed treatment
}

* The format (time-ordering) of data is the same as required by the stremr() function: (I(t), CENS(t), TRT(t), MONITOR(t)).
  MONITOR(t) at time-point t is defined as the indicator of being observed (having an office visit) at time point t+1 (next timepoint after t)
  It is assumed that imp.I(t) is always 0 for the very first time-point.
}
\examples{

\dontrun{
theta <- seq(7,8.5,by=0.5)
FOLLOW.D.DT <- follow.rule.d.DT(data = data, theta = theta,
                 ID = "StudyID", t = "X_intnum", I = "X_a1c",
                 TRT = "X_exposure", CENS = "X_censor", MONITOR = "N.t",
                 rule.names = paste0("new.d",theta))
}
}

