% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataStorageClass.R
\name{get_data}
\alias{get_data}
\title{Extract input data from DataStorageClass}
\usage{
get_data(data)
}
\arguments{
\item{data}{Object of class \code{DataStorageClass} (returned by calling \code{importData} function).}
}
\description{
Extract input data from DataStorageClass
}
