% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_routines.R
\name{get_FUPtimes}
\alias{get_FUPtimes}
\title{Follow-up times by regimen}
\usage{
get_FUPtimes(wts_data, IDnode, tnode)
}
\arguments{
\item{wts_data}{Either a list of data.table containing weights (one for each separate regimen/intervention) or a single data.table with
weights for one regimen / intervention.}

\item{IDnode}{Name of the column containing subject-specific identifier in the input data.}

\item{tnode}{Name of the column containing the time/period variable in the input data.}
}
\value{
A \code{data.table} of subject specific follow-up times for each regimen in \code{wts_data}.
}
\description{
Subject specific follow-up times for each regimen in \code{wts_data}.
}
\seealso{
\code{\link{getIPWeights}} for evaluation of IP-weights.
}
