% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_routines.R
\name{get_MSM_RDs}
\alias{get_MSM_RDs}
\title{Risk Difference Estimates and SEs for IPW-MSM}
\usage{
get_MSM_RDs(MSM, t.periods.RDs, getSEs = TRUE)
}
\arguments{
\item{MSM}{Object returned by \code{\link{survMSM}}.}

\item{t.periods.RDs}{Vector of time-points for evaluation of pairwise risk differences.}

\item{getSEs}{Evaluate the influence curve based RD estimates of standard errors (SEs) along with point estimates?}
}
\value{
...
}
\description{
Returns a table with pair-wise evaluation of risk differences and the corresponding SEs for IPW-MSM output.
}
\seealso{
\code{\link{survMSM}} for estimation with MSM.
}

