% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_routines.R
\name{get_RDs}
\alias{get_RDs}
\title{Produce a long format dataset of risk differences}
\usage{
get_RDs(St_data, St_name, getSEs = TRUE)
}
\arguments{
\item{St_data}{A list containing \code{data.table} survival estimates for each regimen to be contrasted.}

\item{St_name}{The name of the column containing the survival estimates.}

\item{getSEs}{Should standard errors of risk difference also be estimates? Note that
the estimates will only be available when the data.table for each regimen contains a
column of subject-specific influence curve (IC) based estimates. The
column containing these ICs should be named \code{"IC.St"}.}
}
\description{
Produce a long format dataset of risk differences
}

