% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_routines.R
\name{get_TMLE_RDs}
\alias{get_TMLE_RDs}
\title{Risk Difference Estimates and SEs for a list of TMLE outputs}
\usage{
get_TMLE_RDs(TMLE_list, t.periods.RDs)
}
\arguments{
\item{TMLE_list}{A list of objects returned by \code{\link{fit_iterTMLE}}, \code{\link{fit_TMLE}} or \code{\link{fit_GCOMP}}.}

\item{t.periods.RDs}{Vector of time-points for evaluation of pairwise risk differences.}
}
\value{
A list with RD tables. One table for each time-point in \code{t.periods.RDs}.
}
\description{
Produces table(s) with pair-wise risk differences for all regimens that were used for fitting TMLE.
The corresponding SEs are evaluated based on the estimated influence curves (IC).
}
\seealso{
\code{\link{survMSM}} for estimation with MSM.
}
