% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggsurv.R
\name{ggsurv}
\alias{ggsurv}
\title{Survival curves with ggplot2}
\usage{
ggsurv(estimates, CI = TRUE, CI.step = TRUE, plot.cens = TRUE,
  surv.col = "gg.def", cens.col = "gg.def", lty.est = 1, lty.ci = 2,
  size.est = 0.5, size.ci = size.est, cens.size = 2, cens.shape = 3,
  back.white = TRUE, xlab = "Time", ylab = "Survival", main = "",
  legend.position = "right", surv.name = "St." \%+\% attr(estimates[[1]],
  "estimator_short"), SE.name = "SE." \%+\% attr(estimates[[1]],
  "estimator_short"), order.legend = TRUE, t_int_sel = NULL, ...)
}
\arguments{
\item{estimates}{A list, one item per regime / intervention. Each list item must be a data.frame containing the
survival estimates by time for a single regime / intervention.}

\item{CI}{should a 95% confidence interval be plotted? Defaults to \code{TRUE}.
Uses the standard error estimates provided as a separate column of the input data.}

\item{CI.step}{When \code{TRUE} the CIs will be plotted as a step function (same as main plot type).
When \code{FALSE} the CIs are plotted using \code{ggplot2::geom_ribbon}.}

\item{plot.cens}{mark the censored observations?}

\item{surv.col}{colour of the survival estimate. Defaults to black for
one stratum, and to the default \code{ggplot2} colours for multiple
strata. Length of vector with colour names should be either 1 or equal
to the number of strata.}

\item{cens.col}{colour of the points that mark censored observations.}

\item{lty.est}{linetype of the survival curve(s). Vector length should be
either 1 or equal to the number of strata.}

\item{lty.ci}{linetype of the bounds that mark the 95\% CI.}

\item{size.est}{line width of the survival curve}

\item{size.ci}{line width of the 95\% CI}

\item{cens.size}{point size of the censoring points}

\item{cens.shape}{shape of the points that mark censored observations.}

\item{back.white}{if \code{TRUE} the background will not be the default
grey of \code{ggplot2}, but will be white with borders around the plot. Defaults to \code{TRUE}}

\item{xlab}{the label of the x-axis.}

\item{ylab}{the label of the y-axis.}

\item{main}{the plot label.}

\item{legend.position}{Either the coordinates of the legend position inside the plot
(e.g., (0.9, 0.2)) or
the character word denoting the legend orientation with respect to the plot
(e.g., "bottom", "right" or "left").}

\item{surv.name}{The name of the column containing the survival estimates.}

\item{SE.name}{The name of the column containing the standard errors (SE) for each time-point estimate of survival.}

\item{order.legend}{Set to \code{TRUE} to order the legend display by final
survival time (highest first).}

\item{t_int_sel}{The subset of time-point indices for which survival should be plotted.}

\item{...}{Additional arguments (not used).}
}
\value{
An object of class \code{ggplot}
}
\description{
This function produces Kaplan-Meier plots using \code{ggplot2}.
As a first argument it needs a \code{survfit} object, created by the
\code{survival} package.
}
\author{
Original R code by Edwin Thoen \email{edwinthoen@gmail.com}, modified by Oleg Sofrygin \email{oleg.sofrygin@gmail.com}
}

