% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SL.R
\name{h2o.glm_nn}
\alias{h2o.glm_nn}
\title{h2o glm wrapper for non-negative least squares}
\usage{
h2o.glm_nn(..., non_negative = TRUE)
}
\arguments{
\item{...}{All arguments which will be passed down to \code{h2oEnsemble::h2o.glm.wrapper(...)}}

\item{non_negative}{Flag enabling non-negative least squares}
}
\value{
...
}
\description{
This wrapper is used as a default metalearner for SuperLearner / h2oEnsemble.
}

