% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_estimation.R
\name{importData}
\alias{importData}
\title{Import data, define various nodes, define dummies for factor columns and define OData R6 object}
\usage{
importData(data, ID = "Subject_ID", t_name = "time_period", covars,
  CENS = "C", TRT = "A", MONITOR = "N", OUTCOME = "Y", noCENScat = 0L,
  verbose = getOption("stremr.verbose"))
}
\arguments{
\item{data}{Input dataset, can be a data.frame or a data.table}

\item{ID}{...}

\item{t_name}{...}

\item{covars}{...}

\item{CENS}{...}

\item{TRT}{...}

\item{MONITOR}{...}

\item{OUTCOME}{...}

\item{noCENScat}{...}

\item{verbose}{Set to \code{TRUE} to print messages on status and information to the console. Turn this on by default using \code{options(stremr.verbose=TRUE)}.}
}
\value{
...
}
\description{
Import data, define various nodes, define dummies for factor columns and define OData R6 object
}

