% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TMLE_updaters.R
\name{iTMLE.updater.xgb.delta1}
\alias{iTMLE.updater.xgb.delta1}
\alias{iTMLE.updater.xgb.delta2}
\alias{iTMLE.updater.xgb.delta3}
\alias{iTMLE.updater.xgb.delta4}
\alias{iTMLE.updater.xgb}
\alias{predict.iTMLE.updater.xgb}
\title{iTMLE gbm learner / updater}
\usage{
iTMLE.updater.xgb.delta1(Y, X, newX, family, obsWeights, params, ...)

iTMLE.updater.xgb.delta2(Y, X, newX, family, obsWeights, params, ...)

iTMLE.updater.xgb.delta3(Y, X, newX, family, obsWeights, params, ...)

iTMLE.updater.xgb.delta4(Y, X, newX, family, obsWeights, params, ...)

iTMLE.updater.xgb(Y, X, newX, family, obsWeights, params, ...)

\method{predict}{iTMLE.updater.xgb}(object, newdata, offset, ...)
}
\arguments{
\item{Y}{Input outcomes}

\item{X}{Input design matrix with training data.
Must contain a column named "offset", which contains the offsets converted to logit-linear scale.}

\item{newX}{Input design matrix with test data.
Same requirement as for \code{X}: must contain a column named "offset",
which contains the offsets converted to logit-linear scale.}

\item{family}{Link function (ignored).}

\item{obsWeights}{Row-specific weights}

\item{params}{Tuning parameters passed on to \code{xgboost}.}

\item{...}{Additional arguments to be passed on to \code{origami} package.}

\item{object}{Results of calling \code{iTMLE.updater.xgb} functions.}

\item{newdata}{Design matrix with test data for which predictions should be obtained.
May contain a column named "offset", instead of it being passed as a separate argument.}

\item{offset}{Offset (on logit scale if using logistic regression update).}
}
\description{
Performs an SDR update using GBM from \code{xgboost} package with parameter \code{max_delta_step}=1.
This function is passed along as a separate learner
to the SuperLearner implementation of origami package

Performs an SDR update using GBM from \code{xgboost} package with parameter \code{max_delta_step}=2.
This function is passed along as a separate learner
to the SuperLearner implementation of origami package

Performs an SDR update using GBM from \code{xgboost} package with parameter \code{max_delta_step}=3.
This function is passed along as a separate learner
to the SuperLearner implementation of origami package

Performs an SDR update using GBM from \code{xgboost} package with parameter \code{max_delta_step}=4.
This function is passed along as a separate learner
to the SuperLearner implementation of origami package

General GBM learner from \code{xgboost} package.
This function is passed along as a separate learner
to the SuperLearner implementation of origami package
}
