% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TMLE_updaters.R
\name{linear.TMLE.updater.speedglm}
\alias{linear.TMLE.updater.speedglm}
\alias{predict.linear.TMLE.updater}
\title{Linear TMLE glm updater}
\usage{
linear.TMLE.updater.speedglm(Y, X, newX, family, obsWeights, ...)

\method{predict}{linear.TMLE.updater}(object, newdata, offset, ...)
}
\arguments{
\item{Y}{Input outcomes}

\item{X}{Input design matrix with training data.
Must contain a column named "offset", which contains the offsets converted to logit-linear scale.}

\item{newX}{Input design matrix with test data.
Same requirement as for \code{X}: must contain a column named "offset",
which contains the offsets converted to logit-linear scale.}

\item{family}{Link function (ignored).}

\item{obsWeights}{Row-specific weights}

\item{...}{Additional arguments to be passed on to \code{origami} package.}

\item{object}{Results of calling \code{linear.TMLE.updater.speedglm}.}

\item{newdata}{Design matrix with test data for which predictions should be obtained.
May contain a column named "offset", instead of it being passed as a separate argument.}

\item{offset}{Offset (on logit scale if using logistic regression update).}
}
\description{
Performs linear TMLE update using main-terms GLM (family=gaussian()), based on \code{speedglm} package.
This function may be passed to fit_GCOMP / fit_TMLE function.
May be also used as a separate learner for iTMLE.
}
