% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stremr.R
\name{logit_linkinv}
\alias{logit_linkinv}
\title{Fast C++ version of the inverse logit link function (expit)}
\usage{
logit_linkinv(eta)
}
\arguments{
\item{eta}{The argument to be converted}
}
\description{
The conversion performed is given by 1/(1+exp(-eta))
}

