% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stremr-package.R
\docType{data}
\name{O.data.simstudy.g05}
\alias{O.data.simstudy.g05}
\title{An example of a dataset in long format with random monitoring process and no right censoring.}
\format{A data frame with 10,000 observations and variables:
\describe{
  \item{ID}{Unique subject identifier}
  \item{CVD}{Baseline confounder (time invariant)}
  \item{t}{Interger for current time period, range 0-16}
  \item{Y}{Binary outcome}
  \item{lastNat1}{Time since last monitoring event, set to 0 when N[t-1]=0 and then added one for each new period where N[t] is 0.}
  \item{highA1c}{Time-varying confounder}
  \item{TI}{Binary exposure variable}
  \item{C}{Degenerate censoring variable, always set to 0}
  \item{C}{Binary censoring indicator derived from CatC. 0 if CatC is 0 and 1 if CatC is 1 or 2.}
  \item{N}{The random indicator of being monitored (having a visit), simulated as a Bernoulli RV with P(N(t)=1)=0.5}
}}
\usage{
data(O.data.simstudy.g05)
}
\description{
Simulated dataset containing 10,000 i.i.d. observations organized in long format as person-time rows.
The binary exposure is \code{TI} and binary outcome is \code{Y}. See /tests/RUnit_tests_04_simstudyreplica.R
for R code that generated this data as well as R code that uses stremr to analyze this data.
}
\keyword{datasets}

