% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summaryS3methods.R
\name{pander.H2OGrid}
\alias{pander.H2OGrid}
\title{Pander method for H2OGrid S4 class}
\usage{
\method{pander}{H2OGrid}(H2OGridObject)
}
\arguments{
\item{H2OGridObject}{H2OGrid S4 object}
}
\value{
By default this function outputs (see: \code{?cat}) the result.
If you would want to catch the result instead, then call \code{pander_return} instead.
}
\description{
Prints a H2OGrid object in Pandoc's markdown.
}
