% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summaryS3methods.R
\name{pander.H2ORegressionMetrics}
\alias{pander.H2ORegressionMetrics}
\title{Pander method for H2ORegressionMetrics class}
\usage{
\method{pander}{H2ORegressionMetrics}(H2ORegressionMetricsObject)
}
\arguments{
\item{H2ORegressionMetricsObject}{H2ORegressionMetrics object}
}
\value{
By default this function outputs (see: \code{?cat}) the result.
If you would want to catch the result instead, then call \code{pander_return} instead.
}
\description{
Prints a H2ORegressionMetrics object in Pandoc's markdown.
}

