% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BinomialModelClass.R
\name{print.glmfit}
\alias{print.glmfit}
\title{S3 methods for printing model fit summary for glmfit class object}
\usage{
\method{print}{glmfit}(model.fit, ...)
}
\arguments{
\item{model.fit}{The model fit object produced by functions stremr:::glmfit.glm or stremr:::glmfit.speedglm}

\item{...}{Additional options passed on to \code{summary.glmfit}.}
}
\value{
The output is printed with \code{cat}. To capture the markdown-formated model summary use \code{summary.glmfit}.
}
\description{
Prints the modeling summary for the GLM fit (\code{stats::glm.fit} or \code{speedglm::speedglm.wfit})
}

