% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summaryS3methods.R
\name{print.GLMmodel}
\alias{print.GLMmodel}
\title{S3 methods for printing model fit summary for glmfit class object}
\usage{
\method{print}{GLMmodel}(x, ...)
}
\arguments{
\item{x}{The model fit object produced by functions stremr:::fit.glm or stremr:::fit.speedglm}

\item{...}{Additional options passed on to \code{summary.GLMmodel}.}
}
\value{
The output is printed with \code{cat}. To capture the markdown-formated model summary use \code{summary.GLMmodel}.
}
\description{
Prints the modeling summary for the glm fit (\code{stats::glm.fit} or \code{speedglm::speedglm.wfit})
}
