% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BinomialModelClass.R
\name{print.h2ofit}
\alias{print.h2ofit}
\title{S3 methods for printing model fit summary for h2ofit class object}
\usage{
\method{print}{h2ofit}(model.fit, only.coefs = FALSE, ...)
}
\arguments{
\item{model.fit}{The model fit object produced by any stremr S3 function starting with \code{stremr:::h2ofit.}}

\item{only.coefs}{Skip all of the h2o-specific model stats, only print the coefficients table (when running \code{fit.algorithm = "GLM"}).}

\item{...}{Additional options passed on to \code{summary.h2ofit}.}
}
\value{
The output is printed with \code{cat}. To capture the markdown-formated model summary use \code{summary.h2ofit}.
}
\description{
Prints the modeling summary for the h2o model fit (see \code{h2o} R package).
}

