% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Learner_syntax.R
\name{print.ModelStack}
\alias{print.ModelStack}
\title{S3 methods for printing a collection of learners}
\usage{
\method{print}{ModelStack}(x, ...)
}
\arguments{
\item{x}{An object (list) of class ModelStack}

\item{...}{Additional options passed on to \code{print.PredictionModel}.}
}
\description{
Prints the stack models
}
