% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/zzz.R
\name{print_estimtr_opts}
\alias{print_estimtr_opts}
\title{Print Current Option Settings for \code{estimtr}}
\usage{
print_estimtr_opts()
}
\value{
Invisibly returns a list of \code{estimtr} options.
}
\description{
Print Current Option Settings for \code{estimtr}
}
\seealso{
\code{\link{estimtr_options}}
}

