% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_routines.R
\name{print_RDs}
\alias{print_RDs}
\title{Print RD table}
\usage{
print_RDs(RD_tab, dx1, dx2, time, ...)
}
\arguments{
\item{RD_tab}{Risk difference table returned by \code{get_RDs}.}

\item{dx1}{Subset \code{RD_tab} by these \code{dx1} values.}

\item{dx2}{Subset \code{RD_tab} by these \code{dx2} values.}

\item{time}{Subset \code{RD_tab} by these \code{time} values.}

\item{...}{Additional parameters to be passed directly to \code{knitr::kable}.}
}
\value{
A table returned by calling \code{knitr::kable()}.
}
\description{
Print and return RD table in various formats using \code{knitr::kable}
}
