% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summaryS3methods.R
\name{print_tables}
\alias{print_tables}
\alias{print_tables.face.sparse}
\alias{print_tables.brokenstick}
\alias{print_tables.xgb.Booster}
\alias{print_tables.xgb.cv.synchronous}
\alias{print_tables.H2OBinomialModel}
\alias{print_tables.H2ORegressionModel}
\alias{print_tables.GLMmodel}
\title{S3 methods for printing model fit summaries as pander tables}
\usage{
print_tables(model, only.coefs = FALSE, ...)

\method{print_tables}{face.sparse}(model, only.coefs = FALSE, ...)

\method{print_tables}{brokenstick}(model, only.coefs = FALSE, ...)

\method{print_tables}{xgb.Booster}(model, only.coefs = FALSE, ...)

\method{print_tables}{xgb.cv.synchronous}(model, only.coefs = FALSE, ...)

\method{print_tables}{H2OBinomialModel}(model, only.coefs = FALSE, ...)

\method{print_tables}{H2ORegressionModel}(model, only.coefs = FALSE, ...)

\method{print_tables}{GLMmodel}(model, only.coefs = FALSE, ...)
}
\arguments{
\item{model}{The model fit object produced by \code{get_fit} function.}

\item{only.coefs}{Set to \code{TRUE} to only print the table of coefficients (when using \code{fit.algorithm = "glm"}) and
omit all other model-related output statistics.}

\item{...}{Additional options passed on to \code{summary(...)}.}
}
\value{
The output is printed with \code{cat}. To capture the markdown-formated model summary use \code{summary(...)}.
}
\description{
Used internally to prints the modeling summary stats for reporting (see \code{\link{make_model_report}}).
}
