% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataStorageClass.R
\name{remove_g}
\alias{remove_g}
\title{Removing all g model fits from DataStorage object.
Clean up the DataStorage object by removing all g (IPW) modeling fits.
This helps save up some RAM for running G-COMP / TMLE when working with large data.
Call this function AFTER fitting IPW (g), but before calling TMLE.}
\usage{
remove_g(OData, keep_holdout_preds = FALSE)
}
\arguments{
\item{OData}{Data storage object containing the observed data.}

\item{keep_holdout_preds}{Keep the holdout predictions? Needed for CV TMLE only and set to FALSE by default.}
}
\description{
Removing all g model fits from DataStorage object.
Clean up the DataStorage object by removing all g (IPW) modeling fits.
This helps save up some RAM for running G-COMP / TMLE when working with large data.
Call this function AFTER fitting IPW (g), but before calling TMLE.
}
