% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SDR_updaters.R
\name{SDR.updater.glmTMLE}
\alias{SDR.updater.glmTMLE}
\alias{SDR.updater.speedglmTMLE}
\alias{predict.SDR.updater.TMLE}
\title{iTMLE univariate glm learner / updater}
\usage{
SDR.updater.glmTMLE(Y, X, newX, family, obsWeights, ...)

SDR.updater.speedglmTMLE(Y, X, newX, family, obsWeights, ...)

\method{predict}{SDR.updater.TMLE}(object, newdata, ...)
}
\arguments{
\item{Y}{Input outcomes}

\item{X}{Input design matrix with training data.
Must contain a column named "offset", which contains the offsets converted to logit-linear scale.}

\item{newX}{Input design matrix with test data.
Same requirement as for \code{X}: must contain a column named "offset",
which contains the offsets converted to logit-linear scale.}

\item{family}{Link function (ignored).}

\item{obsWeights}{Row-specific weights}

\item{...}{Additional arguments to be passed on to \code{origami} package.}

\item{object}{Results of calling \code{SDR.updater.speedglmTMLE} or \code{SDR.updater.glmTMLE}.}

\item{newdata}{Design matrix with test data for which predictions should be obtained.
Must contain a column named "offset".}
}
\description{
Performs a univariate TMLE update. This function is passed along as a separate learner
to the SuperLearner implementation of origami package

Performs an SDR update using a main-terms GLM (logistic regression with \code{speedglm} package).
This function is passed along as a separate learner
to the SuperLearner implementation of origami package
}
