% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summaryS3methods.R
\name{summary.H2Oensemblemodel}
\alias{summary.H2Oensemblemodel}
\title{S3 methods for getting model fit summary for H2Oensemblemodel class object}
\usage{
\method{summary}{H2Oensemblemodel}(model.fit, only.coefs = FALSE,
  format_table = TRUE, ...)
}
\arguments{
\item{model.fit}{The model fit object produced by any stremr S3 function based on h2o}

\item{only.coefs}{Skip all of the h2o-specific model stats, only print the coefficients table (when running \code{fit.algorithm = "glm"}).}

\item{format_table}{Format the coefficients into a data.frame table (when running \code{fit.algorithm = "glm"})?}

\item{...}{Additional options (not used)}
}
\value{
The markdown-formated model summary returned by \code{pander::pander_return}.
}
\description{
Prints the modeling summary for the h2o model fit (see \code{h2o} R package).
}

