% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summaryS3methods.R
\name{summary.h2ofit}
\alias{summary.h2ofit}
\title{S3 methods for getting model fit summary for h2ofit class object}
\usage{
\method{summary}{h2ofit}(model.fit, only.coefs = FALSE, format_table = TRUE,
  ...)
}
\arguments{
\item{model.fit}{The model fit object produced by any stremr S3 function starting with \code{stremr:::h2ofit.}}

\item{only.coefs}{Skip all of the h2o-specific model stats, only print the coefficients table (when running \code{fit.algorithm = "GLM"}).}

\item{format_table}{Format the coefficients into a data.frame table (when running \code{fit.algorithm = "GLM"})?}

\item{...}{Additional options (not used)}
}
\value{
The markdown-formated model summary returned by \code{pander::pander_return}.
}
\description{
Prints the modeling summary for the h2o model fit (see \code{h2o} R package).
}

