% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summaryS3methods.R
\name{summary.H2ORegressionModel}
\alias{summary.H2ORegressionModel}
\alias{summary.H2OBinomialModel}
\title{S3 methods for fit summary for h2o}
\usage{
\method{summary}{H2ORegressionModel}(object, only.coefs = FALSE, ...)

\method{summary}{H2OBinomialModel}(object, only.coefs = FALSE, ...)
}
\arguments{
\item{object}{The model fit object produced by h2o (and extracted with \code{getmodel_byname}).}

\item{only.coefs}{Skip all of the h2o-specific model stats, only print the coefficients table (when running \code{fit.algorithm = "glm"}).}

\item{...}{Additional options (not used)}
}
\value{
The markdown-formated model summary returned by \code{pander::pander_return}.
}
\description{
Prints the modeling summary for the h2o model fit (see \code{h2o} R package).
}
