% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_estimation.R
\name{survMSM}
\alias{survMSM}
\title{Estimate Survival with a particular MSM for the survival-hazard function using previously fitted weights.}
\usage{
survMSM(wts_data, OData, t_breaks, use_weights = TRUE, stabilize = TRUE,
  trunc_weights = 10^6, weights = NULL, getSEs = TRUE, est_name = "IPW",
  verbose = getOption("stremr.verbose"))
}
\arguments{
\item{wts_data}{A list of \code{data.table}s, each data set is a result of calling the function \code{getIPWeights}. Must contain the treatment/monitoring rule-specific estimated IPTW weights.
This argument can be also a single \code{data.table} obtained with \code{data.table::rbindlist(wts_data)}.}

\item{OData}{The object returned by function \code{fitPropensity}. Contains the input dat and the previously fitted propensity score models for the exposure, monitoring and
right-censoring.}

\item{t_breaks}{The vector of integer (or numeric) breaks that defines the dummy indicators of the follow-up in the observed data. Used for fitting the parametric (or saturated) MSM for the survival hazard. See Details.}

\item{use_weights}{Logical value. Set to \code{FALSE} to ignore the weights in \code{wts_data} and fit a "crude" MSM that does not adjust for the possible confounding due to non-random
assignment of the exposure/censoring and monitoring indicators.}

\item{stabilize}{Set to \code{TRUE} for weight stabilization}

\item{trunc_weights}{Specify the numeric weight truncation value. All final weights exceeding the value in \code{trunc_weights} will be truncated.}

\item{weights}{Optional \code{data.table} with additional observation-time-specific weights.  Must contain columns \code{ID}, \code{t} and \code{weight}.
The column named \code{weight} is merged back into the original data according to (\code{ID}, \code{t}).}

\item{getSEs}{A logical indicator. Set to \code{TRUE} to evaluate the standard errors for the estimated survival by using the MSM influence curve.}

\item{est_name}{A string naming the current MSM estimator. Ignored by the current routine but is used when generating reports with \code{make_report_rmd}.}

\item{verbose}{Set to \code{TRUE} to print messages on status and information to the console. Turn this on by default using \code{options(stremr.verbose=TRUE)}.}
}
\value{
A named list with items containing the MSM estimation results:
 \itemize{
 \item \code{est_name} - .
 \item \code{St} - .
 \item \code{ht} - .
 \item \code{MSM.fit} - .
 \item \code{MSM.intervals} - .
 \item \code{IC.Var.S.d} - .
 \item \code{nID} - .
 \item \code{wts_data} - .
 \item \code{use_weights} - .
 \item \code{trunc_weights} - .
}
}
\description{
Estimate the causal survival curve for a particular stochastic, dynamic or static intervention on the treatment/exposure and monitoring processes based on
the user-specified Marginal Structural Model (MSM) for the counterfactual survival function.
}
\details{
This routine will run the weighted logistic regression using the (possibly-weighted) outcomes from many regimens, with dummy indicators for each treatment/monitoring
regimen available in \code{wts_data} and each follow-up time interval specified in \code{t_breaks}.
When \code{use_weights = TRUE}, the logistic regression for the survival hazard is weighted by the \strong{IPW} (Inverse Probability-Weighted or Horvitz-Thompson) estimated weights
in \code{wts_data}. These IPW weights are based on the previously fitted propensity scores (function \code{fitPropensity}), allowing
adjustment for confounding by possibly non-random assignment to exposure and monitoring and possibly informative right-censoring.
}
\section{Details}{

**********************************************************************

\code{t_breaks} is used for defining the time-intervals of the MSM coefficients for estimation of the survival hazard function.
The first value in \code{t_breaks} defines a dummy variable (indicator) for a fully closed interval, with each subsequent value in \code{t_breaks} defining a single right-closed time-interval.
For example, \code{t_breaks = c(0,1)} will define the MSM dummy indicators: I(min(t) <= t <=0 ), I(0 < t <= 1) and I(1 < t <= max(t)).
On the other hand \code{t_breaks = c(1)} will define the following (more parametric) MSM dummy indicators: I(min(t) <= t <=1 ) and I(1 < t <= max(t)).
If omitted, the default is to define a saturated (non-parametric) MSM with a separate dummy variable for every unique period in the observed data.
}

\section{MSM for the hazard}{

**********************************************************************
}
\examples{
#-------------------------------------------------------------------
# EXAMPLE BASED ON SIMULATED DATA
#-------------------------------------------------------------------
library("data.table")
library("magrittr")
data(OdataCatCENS)
OdataDT <- as.data.table(OdataCatCENS, key=c(ID, t))
# Indicator that the person has never been treated in the past:
OdataDT[, "barTIm1eq0" := as.integer(c(0, cumsum(TI)[-.N]) \%in\% 0), by = ID]

#-------------------------------------------------------------------
# Regressions for modeling the exposure (TRT)
#-------------------------------------------------------------------
gform_TRT <- "TI ~ CVD + highA1c + N.tminus1"
# Fit a separate model for TRT (stratify) for each of the following subsets:
stratify_TRT <- list(
  TI=c(
       # MODEL TI AT t=0
       "t == 0L",
       # MODEL TRT INITATION WHEN MONITORED
       "(t > 0L) & (N.tminus1 == 1L) & (barTIm1eq0 == 1L)",
       # MODEL TRT INITATION WHEN NOT MONITORED
       "(t > 0L) & (N.tminus1 == 0L) & (barTIm1eq0 == 1L)",
       # MODEL TRT CONTINUATION (BOTH MONITORED AND NOT MONITORED)
       "(t > 0L) & (barTIm1eq0 == 1L)"
      ))

#-------------------------------------------------------------------
# Regressions for modeling the categorical censoring (CENS)
#-------------------------------------------------------------------
gform_CENS <- c("CatC ~ highA1c")
# stratify by time-points (separate model for all t<16 and t=16)
stratify_CENS <- list(CatC=c("t < 16", "t == 16"))

#-------------------------------------------------------------------
# Regressions for modeling the monitoring regimen (MONITOR)
#-------------------------------------------------------------------
# Intercept only model, pooling across all time points t
gform_MONITOR <- "N ~ 1"

#-------------------------------------------------------------------
# Define the counterfactual monitoring regimen of interest
#-------------------------------------------------------------------
p <- 0.1 # probability of being monitored at each t is 0.1
OdataDT[, "gstar.N" := ifelse(N == 1L, eval(p), 1-eval(p))]

# Define rule followers/non-followers for two rules: dlow & dhigh
res <- follow.rule.d.DT(OdataDT,
        theta = c(0,1), ID = "ID", t = "t", I = "highA1c",
        CENS = "CatC", TRT = "TI", MONITOR = "N",
        rule.names = c("dlow", "dhigh")) \%>\%
# Merge rule definitions into main dataset:
  merge(OdataDT, ., by=c("ID", "t")) \%>\%
# Estimate hazard and survival for a rule "dhigh":
}
\seealso{
\code{\link{stremr-package}} for the general overview of the package,
}

