% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TMLE_updaters.R
\name{TMLE.updater.NULL}
\alias{TMLE.updater.NULL}
\title{iTMLE NULL learner / updater}
\usage{
TMLE.updater.NULL(Y, X, newX, family, obsWeights, ...)
}
\arguments{
\item{Y}{Input outcomes}

\item{X}{Input design matrix with training data.
Must contain a column named "offset", which contains the offsets converted to logit-linear scale.}

\item{newX}{Input design matrix with test data.
Same requirement as for \code{X}: must contain a column named "offset",
which contains the offsets converted to logit-linear scale.}

\item{family}{Link function (ignored).}

\item{obsWeights}{Row-specific weights}

\item{...}{Additional arguments to be passed on to \code{origami} package.}
}
\description{
Returns the inputs Y as they are, without updating. This function is passed along as a separate learner
to the SuperLearner implementation of origami package
}
