% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataStorageClass.R
\name{trim_rows_after_tmax}
\alias{trim_rows_after_tmax}
\title{Removing not needed rows from observed data (and weights data).
Clean up the DataStorage object by removing all extra rows after certain time-point specified
by \code{tmax}. This will also have the effect of removing all the extra weights.
Can be useful when running TMLE with very large data, trimming the extra rows will free up some RAM.}
\usage{
trim_rows_after_tmax(OData, tmax = OData$max.t)
}
\arguments{
\item{OData}{Data storage object containing the observed data.}

\item{tmax}{Time-point value (max \code{tvals} in TMLE) after which all rows are removed.}
}
\description{
Removing not needed rows from observed data (and weights data).
Clean up the DataStorage object by removing all extra rows after certain time-point specified
by \code{tmax}. This will also have the effect of removing all the extra weights.
Can be useful when running TMLE with very large data, trimming the extra rows will free up some RAM.
}
