% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_estimation.R
\name{define_single_regression}
\alias{define_single_regression}
\title{Define regression models}
\usage{
define_single_regression(OData, regforms, stratify = NULL, params = list())
}
\arguments{
\item{OData}{OData Input data object created by \code{importData} function.}

\item{regforms}{Regression formula, only main terms are allowed.}

\item{stratify}{Expression(s) for creating strata(s) (model will be fit separately on each strata)}

\item{params}{Additional modeling controls (for downstream modeling algorithms)}
}
\description{
Alternative approach to manually define parts of the propsensity score model with formula/strata and model controls
}

