% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_estimation.R
\name{define_single_regression}
\alias{define_single_regression}
\title{Directly specify a single regression model}
\usage{
define_single_regression(
  OData,
  regforms,
  stratify = NULL,
  models = NULL,
  fit_method = stremrOptions("fit_method"),
  fold_column = stremrOptions("fold_column"),
  ...
)
}
\arguments{
\item{OData}{OData Input data object created by \code{importData} function.}

\item{regforms}{Regression formula, only main terms are allowed.}

\item{stratify}{Expression(s) for creating strata(s) (model will be fit separately on each strata)}

\item{models}{Optional parameter specifying the models with \code{gridisl} R package.
Must be an object of class \code{ModelStack} specified with \code{gridisl::defModel} function.}

\item{fit_method}{Model selection approach. Can be \code{"none"} - no model selection,
\code{"cv"} - V fold cross-validation that selects the best model according to lowest cross-validated MSE
(must specify the column name that contains the fold IDs).}

\item{fold_column}{The column name in the input data (ordered factor) that contains the fold IDs to be used as part of the validation sample.
Use the provided function \code{\link{define_CVfolds}} to
define such folds or define the folds using your own method.}

\item{...}{Additional parameters that will be passed down directly to the modeling function.}
}
\description{
This function provides an alternative way to manually define parts of the propsensity score model
with formula/strata and model controls.
This function is for advanced users. It provides explicit and manual control
over every single model fit, e.g., every strata of the exposure propensity score.
}
