% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/estimtr-package.r, R/estimtr.R
\docType{package}
\name{estimtr}
\alias{estimtr}
\alias{estimtr-package}
\title{Estimate the Survival of Intervention on Exposures and MONITORing Process for Right Censored Longitudinal Data.}
\usage{
estimtr(data, ID = "Subj_ID", t = "time_period", covars, CENS = "C",
  TRT = "A", MONITOR = "N", outcome = "Y", gform.CENS, gform.TRT,
  gform.MONITOR, noCENS.cat = 0L, stratify.CENS = NULL,
  stratify.TRT = NULL, stratify.MONITOR = NULL, verbose = FALSE)
}
\arguments{
\item{data}{Observed data, a \code{data.frame} with named columns, containing the baseline covariates,
exposures (\code{Anodes}), the outcomes (\code{Ynode}) and possibly the network column (\code{NETIDnode}), where
network is specified by a vector of strings of friend IDs, each string using \code{optPars$sep} character to separate different friend IDs
(default is \code{optPars$sep=' '}).}

\item{ID}{Unique subject identifier variable in the input data.}

\item{CENS}{The name(s) of the CENSoring variable or variables.
Each separate variable specified in \code{CENS} can be either binary (0/1 valued integer) or categorical (integer).
For binary indicators of CENSoring, the value of 1 indicates the CENSoring or end of follow-up event (this cannot be changed).
For categorical CENSoring variables, by default the value of 0 indicates no CENSoring / continuation of follow-up and other
values indicate different reasons for CENSoring.
Use the argument \code{noCENS.cat} to change the reference (continuation of follow-up) category from default 0 to any other value.
(NOTE: Changing \code{noCENS.cat} has zero effect on coding of the binary CENSoring variables, those have to always use 1 to code the CENSoring event).
Note that factors are not allowed in \code{CENS}.}

\item{gform.CENS}{Regression formula(s) for estimating the propensity score for the censoring mechanism: P(C(t) | W). See Details.}

\item{gform.TRT}{Regression formula(s) for propensity score for the exposure/treatment(s): P(A(t) | W). See Details.}

\item{gform.MONITOR}{Regression formula(s) for estimating the propensity score for the MONITORing process: P(N(t) | W). See Details.
(the observed exposure mechanism), When omitted the regression is defined by \code{sA~sW}, where \code{sA}}

\item{noCENS.cat}{Defaults to 0. Use this to modify the reference category (no CENSoring / continuation of follow-up)
for categorical CENSoring variables specifed in \code{CENS}.}

\item{verbose}{Set to \code{TRUE} to print messages on status and information to the console. Turn this on by default using \code{options(estimtr.verbose=TRUE)}.}

\item{Ynode}{Outcome variable name (column name in \code{data}), assumed normalized between 0 and 1. This can instead be specified
on the left-side of the regression formula in argument \code{Qform}.}

\item{optPars}{A named list of additional optional parameters to be passed to \code{estimtr}, such as
 \code{alpha}, \code{lbound}, \code{family}, \code{YnodeDET},
 \code{h_g0_SummariesModel} and \code{h_gstar_SummariesModel}. See Details below for the description of each parameter.}
}
\value{
A named list with 3 items containing the estimation results for:
 \itemize{
 \item \code{EY_gstar1} - estimates of the mean counterfactual outcome under (stochastic) intervention function \code{f_gstar1} \eqn{(E_{g^*_1}[Y])}.
 \item \code{EY_gstar2} - estimates of the mean counterfactual outcome under (stochastic) intervention function \code{f_gstar2} \eqn{(E_{g^*_2}[Y])},
   \code{NULL} if \code{f_gstar2} not specified.
 \item \code{ATE} - additive treatment effect (\eqn{E_{g^*_1}[Y]} - \eqn{E_{g^*_2}[Y]}) under interventions \code{f_gstar1}
   vs. in \code{f_gstar2}, \code{NULL} if \code{f_gstar2} not specified.
}

Each list item above is itself a list containing the items:
 \itemize{
 \item \code{estimates} - various estimates of the target parameter (network population counterfactual mean under
   (stochastic) intervention).
 \item \code{vars} - the asymptotic variance estimates, for \strong{IPTW} and \strong{TMLE}.
 \item \code{CIs} - CI estimates at \code{alpha} level, for \strong{IPTW} and \strong{TMLE}.
 \item \code{other.vars} - Placeholder for future versions.
 \item \code{h_g0_SummariesModel} - The model fits for P(\code{sA}|\code{sW}) under observed exposure mechanism
   \code{g0}. This is an object of \code{SummariesModel} \pkg{R6} class.
 \item \code{h_gstar_SummariesModel} - The model fits for P(\code{sA}|\code{sW}) under intervention \code{f_gstar1}
   or \code{f_gstar2}. This is an object of \code{SummariesModel} \pkg{R6} class.
}

Currently implemented estimators are:
 \itemize{
 \item \code{h_iptw} - Efficient IPTW based on weights h_gstar/h_gN.
}
}
\description{
The \pkg{estimtr} R package is a tool for estimation of causal survival curve under various user-specified interventions
(e.g., static, dynamic, deterministic, or stochastic).
In particular, the interventions may represent exposures to treatment regimens, the occurrence or non-occurrence of right-censoring
events, or of clinical monitoring events. \pkg{estimtr} enables estimation of a selected set of the user-specified causal quantities of interest,
such as, treatment-specific survival curves and the average risk difference over time.

Estimate the causal survival curve for a particular stochastic, dynamic or static intervention on the treatment exposure regime and  among dependent units with known network structure (in presence of
 interference and/or spillover) using \strong{TMLE} (targeted maximum likelihood estimation), \strong{IPTW}
 (Horvitz-Thompson or the inverse-probability-of-treatment).
}
\section{Documentation}{

\itemize{
\item To see the package vignette use: \code{vignette("estimtr_vignette", package="estimtr")}
\item To see all available package documentation use: \code{help(package = 'estimtr')}
}
}

\section{Routines}{

The following routines will be generally invoked by a user, in the same order as presented below.
\describe{
\item{\code{\link{estimtr}}}{One function for performing estimation}
}
}

\section{Data structures}{

The following most common types of output are produced by the package:
\itemize{
\item \emph{observed data} - input data.frame in long format (repeated measures over time).
}
}

\section{Updates}{

Check for updates and report bugs at \url{http://github.com/osofr/estimtr}.
}

\section{Details}{


The regression formalas in \code{Qform}, \code{hform.g0} and \code{hform.gstar} can include any summary measures names defined in
 \code{sW} and \code{sA}, referenced by their individual variable names or by their aggregate summary measure names.
 For example, \code{hform.g0 = "netA ~ netW"} is equivalent to
 \code{hform.g0 = "A + A_netF1 + A_netF2 ~ W + W_netF1 + W_netF2"} for \code{sW,sA} summary measures defined by
 \code{def_sW(netW=W[[0:2]])} and \code{def_sA(netA=A[[0:2]])}.
}

\section{Additional parameters}{


Some of the parameters that control the estimation in \code{estimtr} can be set by calling the function
 \code{\link{estimtr_options}}.

Additional parameters can be also specified as a named list \code{optPars} argument of the
\code{estimtr} function. The items that can be specified in \code{optPars} are:
\itemize{

\item \code{alpha} - alpha-level for CI calculation (0.05 for 95% CIs);

\item \code{lbound} - One value for symmetrical bounds on P(sW | sW).

\item \code{family} - Family specification for regression models, defaults to binomial (CURRENTLY ONLY BINOMIAL
 FAMILY IS IMPLEMENTED).

\item \code{h_g0_SummariesModel} - Previously fitted model for P(\code{sA}|\code{sW}) under observed exposure mechanism \code{g0},
   returned by the previous runs of the \code{estimtr} function.
   This has to be an object of \code{SummariesModel} \pkg{R6} class. When this argument is specified, all predictions
   P(\code{sA}=\code{sa}|\code{sW}=\code{sw}) under \code{g0} will be based on the model fits provided by this argument.

\item \code{h_gstar_SummariesModel} - Previously fitted model for P(\code{sA}|\code{sW}) under (stochastic) intervention
   specified by \code{f_gstar1} or \code{f_gstar2}. Also an object of \code{SummariesModel} \pkg{R6} class.
   When this argument is specified, the predictions P(\code{sA}=\code{sa}|\code{sW}=\code{sw})
   under \code{f_gstar1} or \code{f_gstar2} will be based on the model fits provided by this argument.
}
}

\section{Specifying the counterfactual intervention function (\code{f_gstar1} and \code{optPars$f_gstar2})}{


The functions \code{f_gstar1} and \code{f_gstar2} can only depend on variables specified by the combined matrix
 of summary measures (\code{sW},\code{sA}), which is passed using the argument \code{data}. The functions should
 return a vector of length \code{nrow(data)} of counterfactual treatments for observations in the input data.
}

\section{IPTW estimator}{

**********************************************************************

\itemize{
\item As described in the following section, the first step is to construct an estimator \eqn{P_{g_N}(sA | sW)}
   for the common (in \code{i}) conditional density \eqn{P_{g_0}(sA | sW)} for common (in \code{i}) unit-level summary
   measures (\code{sA},\code{sW}).

\item The same fitting algorithm is applied to construct an estimator \eqn{P_{g^*_N}(sA^* | sW^*)} of the common (in \code{i})
   conditional density \eqn{P_{g^*}(sA^* | sW^*)} for common (in \code{i}) unit-level summary measures (\code{sA^*},\code{sW^*})
   implied by the user-supplied stochastic intervention \code{f_gstar1} or \code{f_gstar2} and the observed distribution of \code{W}.

\item These two density estimators form the basis of the IPTW estimator,
   which is evaluated at the observed N data points \eqn{O_i=(sW_i, sA_i, Y_i), i=1,...,N} and is given by
   \deqn{\psi^{IPTW}_n = \sum_{i=1,...,N}{Y_i \frac{P_{g^*_N}(sA^*=sA_i | sW=sW_i)}{P_{g_N}(sA=sA_i | sW=sW_i)}}.}
}
}
\examples{
#---------------------------------------------------------------------------------------
# EXAMPLE 1: TO BE COMPLETED
#---------------------------------------------------------------------------------------
}
\seealso{
\code{\link{estimtr-package}} for the general overview of the package,
 \code{\link{def_sW}} for defining the summary measures, \code{\link{eval.summaries}} for
 evaluation and validation of the summary measures,
 and \code{\link{df_netKmax2}}/\code{\link{df_netKmax6}}/\code{\link{NetInd_mat_Kmax6}}
 for examples of network datasets.
}

