% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/estimtr.R
\name{estimtr}
\alias{estimtr}
\title{Estimate Survival with Interventions on Exposure and MONITORing Process in Right Censored Longitudinal Data.}
\usage{
estimtr(data, ID = "Subj_ID", t = "time_period", covars, CENS = "C",
  TRT = "A", MONITOR = "N", OUTCOME = "Y", gform.CENS, gform.TRT,
  gform.MONITOR, noCENS.cat = 0L, stratify.CENS = NULL,
  stratify.TRT = NULL, stratify.MONITOR = NULL, gstar.TRT = NULL,
  gstar.MONITOR = NULL, verbose = FALSE, optPars = list())
}
\arguments{
\item{data}{Observed data in long format. Should be a \code{data.frame} with named columns, containing the time-varying covariates (\code{covars}),
the right-censoring event indicator(s) (\code{CENS}), the exposure variable(s) (\code{TRT}), the monitoring process variable(s) (\code{MONITOR})
and the survival OUTCOME variable (\code{OUTCOME}).}

\item{ID}{Unique subject identifier variable in the input data.}

\item{t}{The name of the time/period variable in the input data.}

\item{covars}{Time varying and baseline covariates. This argument does not need to be specified, by default all variables
that are not in \code{ID}, \code{t}, \code{CENS}, \code{TRT}, \code{MONITOR} and \code{OUTCOME} will be considered as covariates.}

\item{CENS}{CENSoring variable(s) in the input data.
Each separate variable specified in \code{CENS} can be either binary (0/1 valued integer) or categorical (integer).
For binary indicators of CENSoring, the value of 1 indicates the CENSoring or end of follow-up event (this cannot be changed).
For categorical CENSoring variables, by default the value of 0 indicates no CENSoring / continuation of follow-up and other
values indicate different reasons for CENSoring.
Use the argument \code{noCENS.cat} to change the reference (continuation of follow-up) category from default 0 to any other value.
(NOTE: Changing \code{noCENS.cat} has zero effect on coding of the binary CENSoring variables, those have to always use 1 to code the CENSoring event).
Note that factors are not allowed in \code{CENS}.}

\item{TRT}{Exposure/treatment variable(s) in input data.}

\item{MONITOR}{Monitoring variable(s) in input data.}

\item{OUTCOME}{Survival OUTCOME variable name (column name in \code{data}).}

\item{gform.CENS}{Regression formula(s) for estimating the propensity score for the censoring mechanism: P(C(t) | W). See Details.}

\item{gform.TRT}{Regression formula(s) for propensity score for the exposure/treatment(s): P(A(t) | W). See Details.}

\item{gform.MONITOR}{Regression formula(s) for estimating the propensity score for the MONITORing process: P(N(t) | W). See Details.
(the observed exposure mechanism), When omitted the regression is defined by \code{sA~sW}, where \code{sA}}

\item{noCENS.cat}{The level (integer) that indicates CONTINUATION OF FOLLOW-UP for ALL censoring variables. Defaults is 0.
Use this to modify the default reference category (no CENSoring / continuation of follow-up)
for variables specifed in \code{CENS}.}

\item{verbose}{Set to \code{TRUE} to print messages on status and information to the console. Turn this on by default using \code{options(estimtr.verbose=TRUE)}.}

\item{optPars}{A named list of additional optional parameters to be passed to \code{estimtr}, such as
 \code{alpha}, \code{lbound}, \code{family}, \code{YnodeDET},
 \code{h_g0_SummariesModel} and \code{h_gstar_SummariesModel}. See Details below for the description of each parameter.}
}
\value{
A named list with 3 items containing the estimation results for:
 \itemize{
 \item \code{EY_gstar1} - estimates of the mean counterfactual OUTCOME under (stochastic) intervention function \code{f_gstar1} \eqn{(E_{g^*_1}[Y])}.
 \item \code{EY_gstar2} - estimates of the mean counterfactual OUTCOME under (stochastic) intervention function \code{f_gstar2} \eqn{(E_{g^*_2}[Y])},
   \code{NULL} if \code{f_gstar2} not specified.
 \item \code{ATE} - additive treatment effect (\eqn{E_{g^*_1}[Y]} - \eqn{E_{g^*_2}[Y]}) under interventions \code{f_gstar1}
   vs. in \code{f_gstar2}, \code{NULL} if \code{f_gstar2} not specified.
}

Each list item above is itself a list containing the items:
 \itemize{
 \item \code{estimates} - various estimates of the target parameter (network population counterfactual mean under
   (stochastic) intervention).
 \item \code{vars} - the asymptotic variance estimates for \strong{IPTW}.
 \item \code{CIs} - CI estimates at \code{alpha} level for \strong{IPTW}.
 \item \code{other.vars} - Placeholder for future versions.
}

Currently implemented estimators are:
 \itemize{
 \item \code{iptw} - IPTW
}
}
\description{
Estimate the causal survival curve for a particular stochastic, dynamic or static intervention on the treatment/exposure and monitoring process.
 Implements the \strong{IPW} (Inverse Probability-Weighted or Horvitz-Thompson) estimator of the discrete survival hazard function which is mapped into survival function.
}
\section{Details}{


The regression formalas in \code{Qform}, \code{hform.g0} and \code{hform.gstar} can include any summary measures names defined in
 \code{sW} and \code{sA}, referenced by their individual variable names or by their aggregate summary measure names.
 For example, \code{hform.g0 = "netA ~ netW"} is equivalent to
 \code{hform.g0 = "A + A_netF1 + A_netF2 ~ W + W_netF1 + W_netF2"} for \code{sW,sA} summary measures defined by
 \code{def_sW(netW=W[[0:2]])} and \code{def_sA(netA=A[[0:2]])}.
}

\section{Additional parameters}{


Some of the parameters that control the estimation in \code{estimtr} can be set by calling the function \code{\link{estimtr_options}}.

Additional parameters can be also specified as a named list \code{optPars} argument of the \code{estimtr} function.
The items that can be specified in \code{optPars} are:
\itemize{

\item \code{alpha} - alpha-level for CI calculation (0.05 for 95% CIs);

\item \code{lbound} - One value for symmetrical bounds on P(sW | sW).

\item \code{family} - Family specification for regression models, defaults to binomial (CURRENTLY ONLY BINOMIAL
 FAMILY IS IMPLEMENTED).
}
}

\section{Specifying the counterfactual intervention function (\code{f_gstar1} and \code{optPars$f_gstar2})}{


The functions \code{f_gstar1} and \code{f_gstar2} can only depend on variables specified by the combined matrix
 of summary measures (\code{sW},\code{sA}), which is passed using the argument \code{data}. The functions should
 return a vector of length \code{nrow(data)} of counterfactual treatments for observations in the input data.
}

\section{IPTW estimator}{

**********************************************************************

\itemize{
\item As described in the following section, the first step is to construct an estimator \eqn{P_{g_N}(A(t) | L(t))}
   for the probability of exposure \eqn{P_{g_0}(A(t) | W(t))}.

\item Based on the user specified stochastic intervention, we can also obtain \eqn{P_{g^*_N}(A^*(t) | L(t) }

\item Combining the two probabilities forms the basis of the IPTW estimator,
   which is evaluated at the observed N data points \eqn{O_i=((L_i(t), A_i(t): t=0,...,K), Y_i), i=1,...,N} and is given by
   \deqn{\psi^{IPTW}_n = \sum_{i=1,...,N}{Y_i \frac{P_{g^*_N}(A^*(t)=A_i(t) | L(t)=L_i(t))}{P_{g_N}(A(t)=A_i(t) | L(t)=L_i(t))}}.}
}
}
\examples{
#---------------------------------------------------------------------------------------
# EXAMPLE 1: TO BE COMPLETED
#---------------------------------------------------------------------------------------
}
\seealso{
\code{\link{estimtr-package}} for the general overview of the package,
}

