% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/estimtr.R
\name{estimtr}
\alias{estimtr}
\title{Estimate Survival with Interventions on Exposure and MONITORing Process in Right Censored Longitudinal Data.}
\usage{
estimtr(data, ID = "Subj_ID", t = "time_period", covars, CENS = "C",
  TRT = "A", MONITOR = "N", OUTCOME = "Y", gform.CENS, gform.TRT,
  gform.MONITOR, noCENS.cat = 0L, stratify.CENS = NULL,
  stratify.TRT = NULL, stratify.MONITOR = NULL, gstar.TRT = NULL,
  gstar.MONITOR = NULL, verbose = FALSE, optPars = list())
}
\arguments{
\item{data}{Observed data in long format. Should be a \code{data.frame} with named columns, containing the time-varying covariates (\code{covars}),
the right-censoring event indicator(s) (\code{CENS}), the exposure variable(s) (\code{TRT}), the monitoring process variable(s) (\code{MONITOR})
and the survival OUTCOME variable (\code{OUTCOME}).}

\item{ID}{Unique subject identifier variable in the input data.}

\item{t}{The name of the time/period variable in the input data.}

\item{covars}{Time varying and baseline covariates. This argument does not need to be specified, by default all variables
that are not in \code{ID}, \code{t}, \code{CENS}, \code{TRT}, \code{MONITOR} and \code{OUTCOME} will be considered as covariates.}

\item{CENS}{CENSoring variable(s) in the input data.
Each separate variable specified in \code{CENS} can be either binary (0/1 valued integer) or categorical (integer).
For binary indicators of CENSoring, the value of 1 indicates the CENSoring or end of follow-up event (this cannot be changed).
For categorical CENSoring variables, by default the value of 0 indicates no CENSoring / continuation of follow-up and other
values indicate different reasons for CENSoring.
Use the argument \code{noCENS.cat} to change the reference (continuation of follow-up) category from default 0 to any other value.
(NOTE: Changing \code{noCENS.cat} has zero effect on coding of the binary CENSoring variables, those have to always use 1 to code the CENSoring event).
Note that factors are not allowed in \code{CENS}.}

\item{TRT}{Exposure/treatment variable(s) in input data.}

\item{MONITOR}{Monitoring variable(s) in input data.}

\item{OUTCOME}{Survival OUTCOME variable name (column name in \code{data}).}

\item{gform.CENS}{Regression formula(s) for estimating the propensity score for the censoring mechanism: P(C(t) | W). See Details.}

\item{gform.TRT}{Regression formula(s) for propensity score for the exposure/treatment(s): P(A(t) | W). See Details.}

\item{gform.MONITOR}{Regression formula(s) for estimating the propensity score for the MONITORing process: P(N(t) | W). See Details.
(the observed exposure mechanism), When omitted the regression is defined by \code{sA~sW}, where \code{sA}}

\item{noCENS.cat}{The level (integer) that indicates CONTINUATION OF FOLLOW-UP for ALL censoring variables. Defaults is 0.
Use this to modify the default reference category (no CENSoring / continuation of follow-up)
for variables specifed in \code{CENS}.}

\item{verbose}{Set to \code{TRUE} to print messages on status and information to the console. Turn this on by default using \code{options(estimtr.verbose=TRUE)}.}

\item{optPars}{A named list of additional optional parameters to be passed to \code{estimtr}, such as
 \code{alpha}, \code{lbound}, \code{family}, \code{YnodeDET},
 \code{h_g0_SummariesModel} and \code{h_gstar_SummariesModel}. See Details below for the description of each parameter.}
}
\value{
A named list with 3 items containing the estimation results for:
 \itemize{
 \item \code{EY_gstar1} - estimates of the mean counterfactual OUTCOME under (stochastic) intervention function \code{f_gstar1} \eqn{(E_{g^*_1}[Y])}.
 \item \code{EY_gstar2} - estimates of the mean counterfactual OUTCOME under (stochastic) intervention function \code{f_gstar2} \eqn{(E_{g^*_2}[Y])},
   \code{NULL} if \code{f_gstar2} not specified.
 \item \code{ATE} - additive treatment effect (\eqn{E_{g^*_1}[Y]} - \eqn{E_{g^*_2}[Y]}) under interventions \code{f_gstar1}
   vs. in \code{f_gstar2}, \code{NULL} if \code{f_gstar2} not specified.
}

Each list item above is itself a list containing the items:
 \itemize{
 \item \code{estimates} - various estimates of the target parameter (network population counterfactual mean under
   (stochastic) intervention).
 \item \code{vars} - the asymptotic variance estimates for \strong{IPTW}.
 \item \code{CIs} - CI estimates at \code{alpha} level for \strong{IPTW}.
 \item \code{other.vars} - Placeholder for future versions.
}

Currently implemented estimators are:
 \itemize{
 \item \code{iptw} - IPTW
}
}
\description{
Estimate the causal survival curve for a particular stochastic, dynamic or static intervention on the treatment/exposure and monitoring process.
 Implements the \strong{IPW} (Inverse Probability-Weighted or Horvitz-Thompson) estimator of the discrete survival hazard function which is mapped into survival function.
}
\section{Details}{


The regression formalas in \code{Qform}, \code{hform.g0} and \code{hform.gstar} can include any summary measures names defined in
 \code{sW} and \code{sA}, referenced by their individual variable names or by their aggregate summary measure names.
 For example, \code{hform.g0 = "netA ~ netW"} is equivalent to
 \code{hform.g0 = "A + A_netF1 + A_netF2 ~ W + W_netF1 + W_netF2"} for \code{sW,sA} summary measures defined by
 \code{def_sW(netW=W[[0:2]])} and \code{def_sA(netA=A[[0:2]])}.
}

\section{Additional parameters}{


Some of the parameters that control the estimation in \code{estimtr} can be set by calling the function \code{\link{estimtr_options}}.

Additional parameters can be also specified as a named list \code{optPars} argument of the \code{estimtr} function.
The items that can be specified in \code{optPars} are:
\itemize{

\item \code{alpha} - alpha-level for CI calculation (0.05 for 95% CIs);

\item \code{lbound} - One value for symmetrical bounds on P(sW | sW).

\item \code{family} - Family specification for regression models, defaults to binomial (CURRENTLY ONLY BINOMIAL
 FAMILY IS IMPLEMENTED).
}
}

\section{Specifying the counterfactual intervention function (\code{f_gstar1} and \code{optPars$f_gstar2})}{


The functions \code{f_gstar1} and \code{f_gstar2} can only depend on variables specified by the combined matrix
 of summary measures (\code{sW},\code{sA}), which is passed using the argument \code{data}. The functions should
 return a vector of length \code{nrow(data)} of counterfactual treatments for observations in the input data.
}

\section{IPTW estimator}{

**********************************************************************

\itemize{
\item As described in the following section, the first step is to construct an estimator \eqn{P_{g_N}(A(t) | L(t))}
   for the probability of exposure \eqn{P_{g_0}(A(t) | W(t))}.

\item Based on the user specified stochastic intervention, we can also obtain \eqn{P_{g^*_N}(A^*(t) | L(t) }

\item Combining the two probabilities forms the basis of the IPTW estimator,
   which is evaluated at the observed N data points \eqn{O_i=((L_i(t), A_i(t): t=0,...,K), Y_i), i=1,...,N} and is given by
   \deqn{\psi^{IPTW}_n = \sum_{i=1,...,N}{Y_i \frac{P_{g^*_N}(A^*(t)=A_i(t) | L(t)=L_i(t))}{P_{g_N}(A(t)=A_i(t) | L(t)=L_i(t))}}.}
}
}
\examples{
#-------------------------------------------------------------------
# EXAMPLE WITH CATEGORICAL CENSORING (3 levels)
#-------------------------------------------------------------------
library("data.table")
library("magrittr")
data(OdataCatCENS)
OdataDT <- as.data.table(OdataCatCENS, key=c(ID, t))
# Indicator that the person has never been treated in the past:
OdataDT[, "barTIm1eq0" := as.integer(c(0, cumsum(TI)[-.N]) \%in\% 0), by = ID]

#-------------------------------------------------------------------
# Regressions for modeling the exposure (TRT)
#-------------------------------------------------------------------
gform.TRT <- "TI ~ CVD + highA1c + N.tminus1"
# Fit a separate model for TRT (stratify) for each of the following subsets:
stratify.TRT <- list(
  TI=c(
       # MODEL TI AT t=0
       "t == 0L",
       # MODEL TRT INITATION WHEN MONITORED
       "(t > 0L) & (N.tminus1 == 1L) & (barTIm1eq0 == 1L)",
       # MODEL TRT INITATION WHEN NOT MONITORED
       "(t > 0L) & (N.tminus1 == 0L) & (barTIm1eq0 == 1L)",
       # MODEL TRT CONTINUATION (BOTH MONITORED AND NOT MONITORED)
       "(t > 0L) & (barTIm1eq0 == 1L)"
      ))

#-------------------------------------------------------------------
# Regressions for modeling the categorical censoring (CENS)
#-------------------------------------------------------------------
gform.CENS <- c("CatC ~ highA1c")
# stratify by time-points (separate model for all t<16 and t=16)
stratify.CENS <- list(CatC=c("t < 16", "t == 16"))

#-------------------------------------------------------------------
# Regressions for modeling the monitoring regimen (MONITOR)
#-------------------------------------------------------------------
# Intercept only model, pooling across all time points t
gform.MONITOR <- "N ~ 1"

#-------------------------------------------------------------------
# Define the counterfactual monitoring regimen of interest
#-------------------------------------------------------------------
p <- 0.1 # probability of being monitored at each t is 0.1
OdataDT[, "gstar.N" := ifelse(N == 1L, eval(p), 1-eval(p))]

# Define rule followers/non-followers for two rules: dlow & dhigh
res <- follow.rule.d.DT(OdataDT,
        theta = c(0,1), ID = "ID", t = "t", I = "highA1c",
        CENS = "CatC", TRT = "TI", MONITOR = "N",
        rule.names = c("dlow", "dhigh")) \%>\%
# Merge rule definitions into main dataset:
  merge(OdataDT, ., by=c("ID", "t")) \%>\%
# Estimate hazard and survival for one rule (dhigh):
  estimtr(gstar.TRT = "dhigh", gstar.MONITOR = "gstar.N",
        ID = "ID", t = "t", covars = c("highA1c", "lastNat1"),
        CENS = "CatC", gform.CENS = gform.CENS, stratify.CENS = stratify.CENS,
        TRT = "TI", gform.TRT = gform.TRT, stratify.TRT = stratify.TRT,
        MONITOR = "N", gform.MONITOR = gform.MONITOR, OUTCOME = "Y")

res$IPW_estimates
res$dataDT

#

# 1   0   454.3149     52317.95 0.008683729 0.9913163 0.9913163
# 2   1  1017.5525     58103.12 0.017512871 0.9824871 0.9739555
# 3   2  1984.8726     65236.63 0.030425737 0.9695743 0.9443222
# 4   3  2574.1207     72427.57 0.035540621 0.9644594 0.9107604
# 5   4  2992.5615     78301.84 0.038218280 0.9617817 0.8759527
# 6   5  2700.6900     82870.70 0.032589204 0.9674108 0.8474061
# 7   6  2723.0310     85716.79 0.031767768 0.9682322 0.8204859
# 8   7  2430.9155     86839.14 0.027993317 0.9720067 0.7975178
# 9   8  2045.0943     86974.15 0.023513817 0.9764862 0.7787651
# 10  9  1792.4145     86574.65 0.020703687 0.9792963 0.7626418
# 11 10  1241.2401     84669.60 0.014659809 0.9853402 0.7514616
# 12 11  1644.1322     84222.47 0.019521300 0.9804787 0.7367921
# 13 12  1162.6926     82402.67 0.014109889 0.9858901 0.7263960
# 14 13  1345.1770     81187.78 0.016568713 0.9834313 0.7143606
# 15 14  1289.4711     79339.03 0.016252669 0.9837473 0.7027503
# 16 15  1547.5597     77454.59 0.019980220 0.9800198 0.6887092
}
\seealso{
\code{\link{estimtr-package}} for the general overview of the package,
}

