% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmle.R
\name{fitIterTMLE}
\alias{fitIterTMLE}
\title{Iterative TMLE wrapper for \code{fitSeqGcomp}}
\usage{
fitIterTMLE(...)
}
\value{
...
}
\description{
Calls \code{fitSeqGcomp} with argument \code{iterTMLE = TRUE}.
}
\seealso{
\code{\link{fitSeqGcomp}}
}

