% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmle.R
\name{fitTMLE}
\alias{fitTMLE}
\title{TMLE wrapper for \code{fitSeqGcomp}}
\usage{
fitTMLE(...)
}
\arguments{
\item{...}{Arguments that will be passed down to the underlying function \code{fitSeqGcomp}}
}
\value{
\code{data.table} with TMLE survival by time
}
\description{
Calls \code{fitSeqGcomp} with argument \code{TMLE = TRUE}.
}
\seealso{
\code{\link{fitSeqGcomp}}
}

