% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_estimation.R
\name{getIPWeights}
\alias{getIPWeights}
\title{Evaluate weights based gstar_TRT, gstar_MONITOR and observed propensity scores g0, the input is modelfits.g0 and OData object}
\usage{
getIPWeights(OData, gstar_TRT = NULL, gstar_MONITOR = NULL,
  rule_name = paste0(c(gstar_TRT, gstar_MONITOR), collapse = ""),
  stabilize = TRUE)
}
\arguments{
\item{OData}{Input data object created by \code{importData} function.}

\item{gstar_TRT}{Column name in \code{data} containing the counterfactual probabilities of following a specific treatment regimen.}

\item{gstar_MONITOR}{Column name in \code{data} containing the counterfactual probabilities of following a specific monitoring regimen.}

\item{rule_name}{Vector of names with time varying and baseline covariates in \code{data}. This argument does not need to be specified, by default all variables
that are not in \code{ID}, \code{t}, \code{CENS}, \code{TRT}, \code{MONITOR} and \code{OUTCOME} will be considered as covariates.}

\item{stabilize}{Set to \code{TRUE} to return stabilized weights}
}
\value{
...
}
\description{
Requires specification of probabilities for regimens of interest (either as rule followers or as counterfactual indicators)
The output is person-specific data with evaluated weights, wts.DT, only observation-times with non-zero weight are kept
Can be one regimen per single run of this block, which are then combined into a list of output datasets with lapply.
Alternative is to allow input with several rules/regimens, which are automatically combined into a list of output datasets.
}

