% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggsurv.R
\name{ggsurv}
\alias{ggsurv}
\title{Survival curves with ggplot2}
\usage{
ggsurv(
  estimates,
  CI = TRUE,
  CI_line = FALSE,
  plot_cens = TRUE,
  surv_col = "gg.def",
  cens_col = "gg.def",
  lty_est = 1,
  shape_est = seq_along(estimates),
  lty_ci = 2,
  size_est = 0.5,
  size_ci = size_est,
  size_pt = size_est + 0.3,
  cens_size = 2,
  cens_shape = 3,
  back_white = TRUE,
  xlab = "Time",
  ylab = "Survival",
  main = "",
  legend_pos = "right",
  surv_name = "St." \%+\% estimates[[1L]][["est_name"]][1L],
  SE_name = "SE." \%+\% estimates[[1L]][["est_name"]][1L],
  order_legend = TRUE,
  t_int_sel = NULL,
  ymin = NULL,
  ...
)
}
\arguments{
\item{estimates}{A list, one item per regime / intervention. Each list item must be a data.frame containing the
survival estimates by time for a single regime / intervention.}

\item{CI}{should a 95\% confidence interval be plotted? Defaults to \code{TRUE}.
Uses the standard error estimates provided as a separate column of the input data.}

\item{CI_line}{When \code{TRUE} the 95\% CIs will be plotted as a line function (same as main plot type).
When \code{FALSE} the 95\% CIs are plotted using \code{ggplot2::geom_ribbon}.}

\item{plot_cens}{mark the censored observations?}

\item{surv_col}{colour of the survival estimate. Defaults to black for
one stratum, and to the default \code{ggplot2} colours for multiple
strata. Length of vector with colour names should be either 1 or equal
to the number of strata.}

\item{cens_col}{colour of the points that mark censored observations.}

\item{lty_est}{linetype of the survival curve(s). Vector length should be
either 1 or equal to the number of strata.}

\item{shape_est}{shape type of the survival point estimates. Vector length should be
either 1 or equal to the number of strata.}

\item{lty_ci}{linetype of the bounds that mark the 95\% CI.}

\item{size_est}{line width of the survival curve}

\item{size_ci}{line width of the 95\% CI}

\item{size_pt}{point size of the survival estimate at each time-point}

\item{cens_size}{point size of the censoring points}

\item{cens_shape}{shape of the points that mark censored observations.}

\item{back_white}{if \code{TRUE} the background will not be the default
grey of \code{ggplot2}, but will be white with borders around the plot. Defaults to \code{TRUE}}

\item{xlab}{the label of the x-axis.}

\item{ylab}{the label of the y-axis.}

\item{main}{the plot label.}

\item{legend_pos}{Either the coordinates of the legend position inside the plot
(e.g., (0.9, 0.2)) or
the character word denoting the legend orientation with respect to the plot
(e.g., "bottom", "right" or "left").}

\item{surv_name}{The name of the column containing the survival estimates.}

\item{SE_name}{The name of the column containing the standard errors (SE) for each time-point estimate of survival.}

\item{order_legend}{Set to \code{TRUE} to order the legend display by final
survival time (highest first).}

\item{t_int_sel}{The subset of time-point indices for which survival should be plotted.}

\item{ymin}{The minimum value of the y axis. The default (\code{ymin=NULL}) is to use \code{ggplot}
to automatically adjust the limits of the y-axis.}

\item{...}{Additional arguments (not used).}
}
\value{
An object of class \code{ggplot}
}
\description{
This function produces plots of the survival point estimates using \code{ggplot2}.
As a first argument it needs a list of estimates created by the one of the estimation function of the \code{stremr} package.
See http://ggobi.github.io/ggally/#ggallyggsurv for additional modifications of the resulting output plot.
}
\author{
Original R code by Edwin Thoen \email{edwinthoen@gmail.com}, modified by Oleg Sofrygin \email{oleg.sofrygin@gmail.com}
}
