% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reports.R
\name{make_report_rmd}
\alias{make_report_rmd}
\title{Generate report(s) with modeling stats and survival estimates using pandoc.}
\usage{
make_report_rmd(
  OData,
  MSM,
  NPMSM,
  TMLE,
  GCOMP,
  WTtables,
  FUPtables,
  MSM.RDtables,
  TMLE.RDtables,
  plotKM = FALSE,
  printEstimateTables = FALSE,
  format = c("html", "pdf", "word"),
  skip.modelfits = FALSE,
  file.name = getOption("stremr.file.name"),
  file.path = getOption("stremr.file.path"),
  openFile = TRUE,
  serve_html_rmote = FALSE,
  keep_md = FALSE,
  keep_tex = FALSE,
  save_report_data = FALSE,
  use_ggplot = TRUE,
  ...
)
}
\arguments{
\item{OData}{Input data object returned by the function \code{\link{importData}}.}

\item{MSM}{The MSM object fits returned by the function \code{\link{survMSM}}.}

\item{NPMSM}{Optional list of a resulting calls to \code{survNPMSM} or a result of a single call to \code{\link{survNPMSM}}.}

\item{TMLE}{Optional list of a resulting calls to \code{fit_TMLE} or a result of a single call to \code{\link{fit_TMLE}}.}

\item{GCOMP}{Optional list of a resulting calls to \code{fit_GCOMP} or a result of a single call to \code{\link{fit_GCOMP}}.}

\item{WTtables}{Table(s) with distribution(s) of the IPTW weights, a result of calling the function \code{\link{get_wtsummary}}}

\item{FUPtables}{Subject-specific \code{data.table} with maximum follow-up time saved for each subject in the column named \code{max.t}.
See \code{\link{get_FUPtimes}} for additional details.}

\item{MSM.RDtables}{List of tables with risk differences returned by the function \code{\link{get_MSM_RDs}}.}

\item{TMLE.RDtables}{List of tables with risk differences returned by the function \code{\link{get_TMLE_RDs}}.}

\item{plotKM}{Logical, set to \code{TRUE} to plot KM survival curves when \code{NPMSM} argument is specified. Default is \code{FALSE}.}

\item{printEstimateTables}{...}

\item{format}{Choose the Pandoc output format for the report file (html, pdf or word).
Note that the html report file is always produced in addition to any other selected format.}

\item{skip.modelfits}{Do not report any of the modeling stats.}

\item{file.name}{File name for the report file without extension. Default file name is assigned based on the current date.}

\item{file.path}{Directory path where the report file(s) should be written. Default is to use the system temporary directory.}

\item{openFile}{Open the report file with OS default viewer?}

\item{serve_html_rmote}{Serve the html report as a webpage via R package "rmote".}

\item{keep_md}{Keep the source .md files?}

\item{keep_tex}{Keep the source .tex files for pdf output?}

\item{save_report_data}{Save the data needed for generating this report as a list in a separate 'report_name.Rd' file.}

\item{use_ggplot}{Set to \code{TRUE} to enable plotting with ggplot.}

\item{...}{Additional arguments may specify the report title (\code{author}), author (\code{title}).
Specifying the logical flag \code{only.coefs=TRUE} disables printing of all h2o-specific model summaries.
Additional set of arguments control the survival plotting, these are passed on to the function \code{f_plot_survest}:
\code{t_int_sel}, \code{y_lab}, \code{x_lab}, \code{miny}, \code{x_legend}, \code{y_legend}.}
}
\value{
String specifying the path to the main report file.
}
\description{
Generate report(s) with modeling stats and survival estimates using pandoc.
}
