% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{print_stremr_opts}
\alias{print_stremr_opts}
\title{Show all current option settings}
\usage{
print_stremr_opts()
}
\value{
Invisibly returns a list of \code{stremr} options.
}
\description{
Show all current option settings
}
\seealso{
\code{\link{set_all_stremr_options}}
}
