% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{set_all_stremr_options}
\alias{set_all_stremr_options}
\title{Setting all possible \code{stremr} options at once.}
\usage{
set_all_stremr_options(
  fit_method = c("none", "cv"),
  fold_column = NULL,
  lower_bound_zero_Q = TRUE,
  skip_update_zero_Q = TRUE,
  up_trunc_offset = 20,
  low_trunc_offset = -10,
  eps_tol = 10^-5
)
}
\arguments{
\item{fit_method}{Specify the default method for model selection.
Possible options are \code{"none"} - no model selection and no cross-validation (when using only a single model, e.g., \code{speedglm__glm})
or \code{"cv"} - perform V-fold cross-validation to select the best model based on lowest MSE.
Note that when code{fit_method = "cv"}, the argument \code{fold_column} also needs to be specified.}

\item{fold_column}{The column name in the input data (ordered factor) that contains the fold IDs to be used as part of the validation sample.
Use the provided function \code{\link{define_CVfolds}} to
define such folds or define the folds using your own method.}

\item{lower_bound_zero_Q}{Set to \code{TRUE} to bound the observation-specific Qs during the TMLE update step away from zero (with minimum value set at 10^-4).
Can help numerically stabilize the TMLE intercept estimates in some small-sample cases. Has no effect when \code{TMLE} = \code{FALSE}.}

\item{skip_update_zero_Q}{Set to \code{FALSE} to perform TMLE update with glm even when all of the Q's are zero.
When set to \code{TRUE} the TMLE update step is skipped if the predicted Q's are either all 0 or near 0, with TMLE intercept being set to 0.}

\item{up_trunc_offset}{The upper bound for the TMLE offset during the TMLE GLM update step.}

\item{low_trunc_offset}{The lower bound for the TMLE offset during the TMLE GLM update step.}

\item{eps_tol}{Used for TMLE GLM update step. Set the tolerance for testing that the outcomes (\code{Qkplus1}) are all 0 or are all 1.}
}
\value{
Invisibly returns a list with old option settings.
}
\description{
Options that control \code{stremr} package.
\strong{Calling this function will reset all unspecified options (omitted arguments) to their default values}!
The preferred way to set options for \code{stremr} is to use \code{\link{stremrOptions}}, which allows specifying individual options without having to reset all other options.
To reset all options to their defaults simply run \code{set_all_stremr_options()} without any parameters/arguments.
}
\seealso{
\code{\link{stremrOptions}}, \code{\link{print_stremr_opts}}
}
