% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summaryS3methods.R
\name{summary.GLMmodel}
\alias{summary.GLMmodel}
\title{S3 methods for fit summary for glmfit class}
\usage{
\method{summary}{GLMmodel}(object, format_table = TRUE, ...)
}
\arguments{
\item{object}{The model fit object produced by functions stremr:::glmfit.glm or stremr:::glmfit.speedglm}

\item{format_table}{Format the coefficients into a data.frame table?}

\item{...}{Additional options (not used)}
}
\value{
The markdown-formated model summary returned by \code{pander::pander_return}.
}
\description{
Prints the modeling summary for the GLM fit (\code{stats::glm.fit} or \code{speedglm::speedglm.wfit})
}
