% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_estimation.R
\name{survNPMSM}
\alias{survNPMSM}
\title{Non-parametric (saturated) MSM for survival based on previously evaluated IP weights.}
\usage{
survNPMSM(wts_data, OData)
}
\arguments{
\item{wts_data}{A list of \code{data.table}s, each data set is a result of calling the function \code{getIPWeights}. Must contain the treatment/monitoring rule-specific estimated IPTW weights.
This argument can be also a single \code{data.table} obtained with \code{data.table::rbindlist(wts_data)}.}

\item{OData}{The object returned by function \code{fitPropensity}. Contains the input dat and the previously fitted propensity score models for the exposure, monitoring and
right-censoring.}
}
\value{
A data.table with hazard and survival function estimates by time. Also include the unadjusted Kaplan-Maier estimates.
}
\description{
Non-parametric (saturated) MSM for survival based on previously evaluated IP weights.
}

