% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TMLE_updaters.R
\name{TMLE.updater.glm}
\alias{TMLE.updater.glm}
\alias{TMLE.updater.speedglm}
\alias{predict.TMLE.updater}
\title{iTMLE univariate glm learner / updater}
\usage{
TMLE.updater.glm(Y, X, newX, family, obsWeights, ...)

TMLE.updater.speedglm(Y, X, newX, family, obsWeights, ...)

\method{predict}{TMLE.updater}(object, newdata, offset, ...)
}
\arguments{
\item{Y}{Input outcomes}

\item{X}{Input design matrix with training data.
Must contain a column named "offset", which contains the offsets converted to logit-linear scale.}

\item{newX}{Input design matrix with test data.
Same requirement as for \code{X}: must contain a column named "offset",
which contains the offsets converted to logit-linear scale.}

\item{family}{Link function (ignored).}

\item{obsWeights}{Row-specific weights}

\item{...}{Additional arguments to be passed on to \code{origami} package.}

\item{object}{Results of calling \code{TMLE.updater.speedglm} or \code{TMLE.updater.glm}.}

\item{newdata}{Design matrix with test data for which predictions should be obtained.
May contain a column named "offset", instead of it being passed as a separate argument.}

\item{offset}{Offset (on logit scale if using logistic regression update).}
}
\description{
Performs a univariate TMLE update. This function is passed along as a separate learner
to the SuperLearner implementation of origami package

Performs TMLE update using a main-terms GLM (logistic regression with \code{speedglm} package).
This function may be passed to fit_GCOMP / fit_TMLE function.
May be also used as a separate learner for iTMLE.
}
